Add test to delete functionality
