--- src/addTodo.test.js
@@ -1,89 +1,0 @@
-import { Todos } from './modules/Todos.js';
-
-describe('Todos', () => {
-    describe('addTask', () => {
-        beforeEach(() => {
-            localStorage.clear();
-        });
-
-        it('should add a new task to the tasks array and save to local storage', () => {
-            const todos = new Todos();
-
-            todos.addTask('Buy milk');
-
-            expect(todos.tasks.length).toBe(1);
-            expect(todos.tasks[0].description).toBe('Buy milk');
-            expect(todos.tasks[0].completed).toBe(false);
-            expect(todos.tasks[0].index).toBe(1);
-
-            const savedTasks = JSON.parse(localStorage.getItem('tasks'));
-            expect(savedTasks.length).toBe(1);
-            expect(savedTasks[0].description).toBe('Buy milk');
-            expect(savedTasks[0].completed).toBe(false);
-            expect(savedTasks[0].index).toBe(1);
-        });
-
-        it('should not add a task if the description is empty', () => {
-            const todos = new Todos();
-
-            todos.addTask('');
-
-            expect(todos.tasks.length).toBe(0);
-            expect(localStorage.getItem('tasks')).toBeNull();
-        });
-    });
-});
-
-
-// describe('Todos', () => {
-//     describe('deleteTask', () => {
-//         beforeEach(() => {
-//             localStorage.clear();
-//         });
-
-//         it('should update the index property of remaining tasks after a task is deleted', () => {
-//             const todos = new Todos();
-//             todos.addTask('Buy milk');
-//             todos.addTask('Do laundry');
-//             todos.addTask('Clean house');
-
-//             todos.deleteTask(2);
-
-//             expect(todos.tasks.length).toBe(2);
-//             expect(todos.tasks[0].description).toBe('Buy milk');
-//             expect(todos.tasks[0].completed).toBe(false);
-//             expect(todos.tasks[0].index).toBe(1);
-//             expect(todos.tasks[1].description).toBe('Clean house');
-//             expect(todos.tasks[1].completed).toBe(false);
-//             expect(todos.tasks[1].index).toBe(2);
-
-//             const savedTasks = JSON.parse(localStorage.getItem('tasks'));
-//             expect(savedTasks.length).toBe(2);
-//             expect(savedTasks[0].description).toBe('Buy milk');
-//             expect(savedTasks[0].completed).toBe(false);
-//             expect(savedTasks[0].index).toBe(1);
-//             expect(savedTasks[1].description).toBe('Clean house');
-//             expect(savedTasks[1].completed).toBe(false);
-//             expect(savedTasks[1].index).toBe(2);
-//         });
-
-
-//         it('should not delete a task if the index is invalid', () => {
-//             const todos = new Todos();
-//             todos.addTask('Buy milk');
-
-//             todos.deleteTask(2);
-
-//             expect(todos.tasks.length).toBe(1);
-//             expect(todos.tasks[0].description).toBe('Buy milk');
-//             expect(todos.tasks[0].completed).toBe(false);
-//             expect(todos.tasks[0].index).toBe(1);
-
-//             const savedTasks = JSON.parse(localStorage.getItem('tasks'));
-//             expect(savedTasks.length).toBe(1);
-//             expect(savedTasks[0].description).toBe('Buy milk');
-//             expect(savedTasks[0].completed).toBe(false);
-//             expect(savedTasks[0].index).toBe(1);
-//         });
-//     });
-// });
